updated comment posting/refactored some profile code
