+++ js/profile.mjs
@@ -1,29 +1,9 @@
-import {initialiseAPIHandler, createBasicPost, createAPost, showInput, newPost, isValidImgLink, createAvatar} from "./modules/main.mjs"
-
-//-------------------Create API handler -----------------------
-const API = initialiseAPIHandler();
+import {addEditProfileListeners, renderProfileContent, createBasicPost, makeAPostListener, showInput, isValidImgLink, createAvatar, API} from "./modules/main.mjs"
 
 //-------------------page grabs-----------------------
-//users banner, avatar and name
-const profileBanner = document.querySelector("#profileBanner");
-const profileImage = document.querySelector("#profileImage");
-const heading = document.querySelector("h1");
 //post comment form
 const postsContainer = document.querySelector("#post-feed");
 const postCommentSection = document.querySelector("#post-comment");
-const postCommentForm = document.querySelector("#post-comment-form");
-const imageBtn = document.querySelector("#img-btn");
-const imageContainer = document.querySelector("#image-container");
-const tagsBtn = document.querySelector("#tags-btn");
-const tagsContainer = document.querySelector("#tags-container");
-//edit user details form
-const aboutContainer = document.querySelector("#about");
-const editContainer = document.querySelector("#edit");
-const editError = document.querySelector("#edit-error");
-const editBtn = document.querySelector("#edit-btn");
-const editProfileForm = document.querySelector("#edit-profile-form");
-const avatarInput = document.querySelector("#avatar");
-const bannerInput = document.querySelector("#banner");
 //followers
 const followersContainer = document.querySelector("#followers");
 const followingContainer = document.querySelector("#following");
@@ -55,87 +35,23 @@ const user = defineUser();
 async function initialiseProfileFunctionality({name, avatar, banner, followers}){
   //is this the logged in users profile
   if(API.name === name){
-    //displays the edit button, targets inputs and assigns the event listener
-    editContainer.classList.remove("hidden");
-    editBtn.addEventListener("click", showProfileForm);
-    //fills inputs with current values in case of only changing 1 input
-    bannerInput.value = banner;
-    avatarInput.value = avatar;
-    editProfileForm.addEventListener("submit", updateProfile);
-
+    //enables profile editing
+    addEditProfileListeners();
     //enables posting to your feed.
     postCommentSection.classList.remove("hidden");
-    imageBtn.addEventListener("click", function(){showInput(imageContainer, 84)});
-    tagsBtn.addEventListener("click", function(){showInput(tagsContainer, 84)});
-    //postCommentForm.addEventListener("submit", postYourComment);
+    makeAPostListener();
   } else{
-    //shows the follow button for your own profile
+    //hides the follow button for your own profile
     followContainer.classList.remove("hidden");
     //gets an array of follower names
     const followerNames = followers.map((follower) => follower.name);
-    //checks if 
+    //checks if your following or not
     if(followerNames.includes(API.name)){
       followBtn.innerText = "Unfollow";
     }
     followBtn.setAttribute("user", name);
     followBtn.addEventListener("click", followUser);
   }
-}
-
-//------------------- Edit Form -----------------------
-/**
- * Shows the edit form.
- */
- function showProfileForm(){
-  showInput(editProfileForm, 210);
-  //editBtn.classList.add("hidden");
-}
-
-/**
- * Takes the edit form and validates the image links.
- * Then posts updates to users profile images if valid.
- * Takes response to update their profile page details.
- * @param {Event} submit submission of the form
- */
-async function updateProfile(submit){
-  submit.preventDefault();
-  if(!isValidImgLink(bannerInput.value)){ editError.innerHTML="Invalid Image Link"}
-  if(!isValidImgLink(avatarInput.value)){ editError.innerHTML="Invalid Image Link"}
-  if(isValidImgLink(avatarInput.value) && isValidImgLink(bannerInput.value)){
-    editError.innerHTML=""
-    const body = JSON.stringify({banner:bannerInput.value, avatar:avatarInput.value});
-    const response = await API.updateProfile(body);
-    renderProfileContent(response);
-    showProfileForm();
-  }
-}
-
-//------------------- New Post Form -----------------------
-
-/**
- * Posts the data from the post form,
- * then
- * @param {Event} submit 
- */
-async function postYourComment(submit){
-  try{
-    submit.preventDefault();
-    const formDate = new FormData(submit.target);
-    const bodyData = Object.fromEntries(formDate.entries());
-    // const title = document.querySelector("#title");
-    // const body = document.querySelector("#body");
-    // const media = document.querySelector("#media");
-    // const tags = document.querySelector("#tags");
-    // const bodyData = new newPost(title.value, body.value, media.value, tags.value);
-    // const response = await postComment(API, JSON.stringify(bodyData.returnBody()))
-    const response = await API.createPost(JSON.stringify(bodyData));
-    console.log(response);
-    postCommentForm.reset();
-    getUsersPosts();
-  } catch(error){
-    console.log(error)
-  }
-
 }
 
 //------------------- Follower Btn Functionality  -----------------------
@@ -159,19 +75,7 @@ async function postYourComment(submit){
   }
 }
 
-
 //------------------- Render and Edit html -----------------------
-
-/**
- * Fills in page content with user info.
- * @param {Object} UserData 
- */
-function renderProfileContent({banner, avatar, name, meta = ""}){
-  profileBanner.src = banner;
-  profileImage.src = avatar;
-  heading.innerText = name;     
-  // aboutContainer.innerText = meta // awaiting extra content???
-}
 
 /**
  * Gets users posts and renders them on the page.
@@ -180,10 +84,6 @@ async function getUsersPosts(){
   const postData = await API.getPosts(); 
   const yourPosts = postData.filter(post => post.author.name === user);
   postsContainer.innerHTML= "";
-  // long form version but using template is easier.
-  // yourPosts.forEach(post => {
-  //   postsContainer.appendChild(createBasicPost(post));
-  // });
   yourPosts.forEach(post => {
   postsContainer.appendChild(createBasicPost(post));
   });
