+++ profile.html
@@ -141,11 +141,18 @@
                       ></textarea>
                       <div class="hidden closing" id="image-container">
                         <label for="media" class="text-white ms-2 pt-2">Image Link</label>
-                        <input type="text" class="form-control rounded-0 border-0 mt-1 bg-tertiary text-white place-text-light" id="media" name="media" placeholder="www.example.com/image.jpg" />
+                        <input
+                          type="text"
+                          class="form-control rounded-0 border-0 mt-1 bg-tertiary text-white place-text-light"
+                          id="media"
+                          name="media"
+                          placeholder="www.example.com/image.jpg"
+                          disabled
+                        />
                       </div>
                       <div class="hidden closing" id="tags-container">
                         <label for="tags" class="text-white ms-2 pt-2">Tags</label>
-                        <input type="text" class="form-control rounded-0 border-0 mt-1 bg-tertiary text-white place-text-light" id="tags" placeholder="#OhSoSocial!" name="tags" />
+                        <input type="text" class="form-control rounded-0 border-0 mt-1 bg-tertiary text-white place-text-light" id="tags" placeholder="#OhSoSocial!" name="tags" disabled />
                       </div>
                     </div>
                     <div class="p-2 d-flex justify-content-between">
